python-service/requirements.txt
flask==2.3.3
openai==0.28.0
pillow==10.0.0
requests==2.31.0
gunicorn==21.2.0